# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface Scenario {
  id: number;
  name: string;
  description: string;
  category: string;
  difficulty: string;
  systemPrompt: string;
  icon: string;
}

export interface PracticeSession {
  id: number;
  scenarioId: number;
  conversationId: number;
  scenarioName: string;
  durationSeconds: number | null;
  feedback: string | null;
  score: number | null;
  createdAt: string;
}

export interface PracticeSessionWithMessages {
  id: number;
  scenarioId: number;
  conversationId: number;
  scenarioName: string;
  durationSeconds: number | null;
  feedback: string | null;
  score: number | null;
  createdAt: string;
  messages: OpenaiMessage[];
  scenario: Scenario;
}

export interface CreateSessionBody {
  scenarioId: number;
}

export interface SessionFeedback {
  feedback: string;
  score: number;
  suggestions: string[];
}

export type PracticeStatsSessionsByCategory = { [key: string]: number };

export interface PracticeStats {
  totalSessions: number;
  totalMinutes: number;
  averageScore: number;
  sessionsByCategory: PracticeStatsSessionsByCategory;
  recentSessions: PracticeSession[];
}