# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: practice
    description: IT English practice scenarios and sessions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiVoiceMessageBody"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}

  /practice/scenarios:
    get:
      operationId: listScenarios
      tags: [practice]
      summary: List all IT English practice scenarios
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"

  /practice/sessions:
    get:
      operationId: listSessions
      tags: [practice]
      summary: List practice sessions for history
      responses:
        "200":
          description: List of past sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeSession"
    post:
      operationId: createSession
      tags: [practice]
      summary: Create a new practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session with linked conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"

  /practice/sessions/{id}:
    get:
      operationId: getSession
      tags: [practice]
      summary: Get a session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSessionWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /practice/sessions/{id}/feedback:
    post:
      operationId: generateFeedback
      tags: [practice]
      summary: Generate AI feedback for a practice session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback for the session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionFeedback"

  /practice/stats:
    get:
      operationId: getPracticeStats
      tags: [practice]
      summary: Get user practice statistics
      responses:
        "200":
          description: Practice stats overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    SendOpenaiVoiceMessageBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Scenario:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        difficulty:
          type: string
        systemPrompt:
          type: string
        icon:
          type: string
      required:
        - id
        - name
        - description
        - category
        - difficulty
        - systemPrompt
        - icon

    PracticeSession:
      type: object
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        conversationId:
          type: integer
        scenarioName:
          type: string
        durationSeconds:
          type: integer
          nullable: true
        feedback:
          type: string
          nullable: true
        score:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scenarioId
        - conversationId
        - scenarioName
        - durationSeconds
        - feedback
        - score
        - createdAt

    PracticeSessionWithMessages:
      type: object
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        conversationId:
          type: integer
        scenarioName:
          type: string
        durationSeconds:
          type: integer
          nullable: true
        feedback:
          type: string
          nullable: true
        score:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
        scenario:
          $ref: "#/components/schemas/Scenario"
      required:
        - id
        - scenarioId
        - conversationId
        - scenarioName
        - durationSeconds
        - feedback
        - score
        - createdAt
        - messages
        - scenario

    CreateSessionBody:
      type: object
      properties:
        scenarioId:
          type: integer
      required:
        - scenarioId

    SessionFeedback:
      type: object
      properties:
        feedback:
          type: string
        score:
          type: integer
        suggestions:
          type: array
          items:
            type: string
      required:
        - feedback
        - score
        - suggestions

    PracticeStats:
      type: object
      properties:
        totalSessions:
          type: integer
        totalMinutes:
          type: integer
        averageScore:
          type: number
        sessionsByCategory:
          type: object
          additionalProperties:
            type: integer
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/PracticeSession"
      required:
        - totalSessions
        - totalMinutes
        - averageScore
        - sessionsByCategory
        - recentSessions