# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Send audio and receive a streaming voice response
 */
export const SendOpenaiVoiceMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiVoiceMessageBody = zod.object({
  audio: zod.string().describe("Base64-encoded audio data"),
});

/**
 * @summary List all IT English practice scenarios
 */
export const ListScenariosResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  difficulty: zod.string(),
  systemPrompt: zod.string(),
  icon: zod.string(),
});
export const ListScenariosResponse = zod.array(ListScenariosResponseItem);

/**
 * @summary List practice sessions for history
 */
export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  scenarioId: zod.number(),
  conversationId: zod.number(),
  scenarioName: zod.string(),
  durationSeconds: zod.number().nullable(),
  feedback: zod.string().nullable(),
  score: zod.number().nullable(),
  createdAt: zod.coerce.date(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a new practice session
 */
export const CreateSessionBody = zod.object({
  scenarioId: zod.number(),
});

/**
 * @summary Get a session with messages
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  scenarioId: zod.number(),
  conversationId: zod.number(),
  scenarioName: zod.string(),
  durationSeconds: zod.number().nullable(),
  feedback: zod.string().nullable(),
  score: zod.number().nullable(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  scenario: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string(),
    category: zod.string(),
    difficulty: zod.string(),
    systemPrompt: zod.string(),
    icon: zod.string(),
  }),
});

/**
 * @summary Generate AI feedback for a practice session
 */
export const GenerateFeedbackParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateFeedbackResponse = zod.object({
  feedback: zod.string(),
  score: zod.number(),
  suggestions: zod.array(zod.string()),
});

/**
 * @summary Get user practice statistics
 */
export const GetPracticeStatsResponse = zod.object({
  totalSessions: zod.number(),
  totalMinutes: zod.number(),
  averageScore: zod.number(),
  sessionsByCategory: zod.record(zod.string(), zod.number()),
  recentSessions: zod.array(
    zod.object({
      id: zod.number(),
      scenarioId: zod.number(),
      conversationId: zod.number(),
      scenarioName: zod.string(),
      durationSeconds: zod.number().nullable(),
      feedback: zod.string().nullable(),
      score: zod.number().nullable(),
      createdAt: zod.coerce.date(),
    }),
  ),
});