# lib/api-zod/src/generated/types/practiceSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeSession {
  id: number;
  scenarioId: number;
  conversationId: number;
  scenarioName: string;
  durationSeconds: number | null;
  feedback: string | null;
  score: number | null;
  createdAt: Date;
}