# lib/api-zod/src/generated/types/practiceSessionWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiMessage } from "./openaiMessage";
import type { Scenario } from "./scenario";

export interface PracticeSessionWithMessages {
  id: number;
  scenarioId: number;
  conversationId: number;
  scenarioName: string;
  durationSeconds: number | null;
  feedback: string | null;
  score: number | null;
  createdAt: Date;
  messages: OpenaiMessage[];
  scenario: Scenario;
}