# lib/api-zod/src/generated/types/practiceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeSession } from "./practiceSession";
import type { PracticeStatsSessionsByCategory } from "./practiceStatsSessionsByCategory";

export interface PracticeStats {
  totalSessions: number;
  totalMinutes: number;
  averageScore: number;
  sessionsByCategory: PracticeStatsSessionsByCategory;
  recentSessions: PracticeSession[];
}