# lib/api-zod/src/generated/types/scenario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Scenario {
  id: number;
  name: string;
  description: string;
  category: string;
  difficulty: string;
  systemPrompt: string;
  icon: string;
}